+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -178,7 +178,7 @@ export type RecipeInterface = {
         userContext: any;
     }): Promise<User | undefined>;
 
-    signInUp(input: {
+    thirdPartySignInUp(input: {
         thirdPartyId: string;
         thirdPartyUserId: string;
         email: {
@@ -194,13 +194,13 @@ export type RecipeInterface = {
           }
     >;
 
-    signUp(input: {
+    emailPasswordSignUp(input: {
         email: string;
         password: string;
         userContext: any;
     }): Promise<{ status: "OK"; user: User } | { status: "EMAIL_ALREADY_EXISTS_ERROR" }>;
 
-    signIn(input: {
+    emailPasswordSignIn(input: {
         email: string;
         password: string;
         userContext: any;
@@ -215,7 +215,17 @@ export type RecipeInterface = {
         token: string;
         newPassword: string;
         userContext: any;
-    }): Promise<{ status: "OK" | "RESET_PASSWORD_INVALID_TOKEN_ERROR" }>;
+    }): Promise<
+        | {
+              status: "OK";
+              /**
+               * The id of the user whose password was reset.
+               * Defined for Core versions 3.9 or later
+               */
+              userId?: string;
+          }
+        | { status: "RESET_PASSWORD_INVALID_TOKEN_ERROR" }
+    >;
 
     updateEmailOrPassword(input: {
         userId: string;
@@ -240,7 +250,7 @@ export type APIInterface = {
               url: string;
           }>);
 
-    emailExistsGET:
+    emailPasswordEmailExistsGET:
         | undefined
         | ((input: {
               email: string;
@@ -274,9 +284,15 @@ export type APIInterface = {
               token: string;
               options: EmailPasswordAPIOptions;
               userContext: any;
-          }) => Promise<{
-              status: "OK" | "RESET_PASSWORD_INVALID_TOKEN_ERROR";
-          }>);
+          }) => Promise<
+              | {
+                    status: "OK";
+                    userId?: string;
+                }
+              | {
+                    status: "RESET_PASSWORD_INVALID_TOKEN_ERROR";
+                }
+          >);
 
     thirdPartySignInUpPOST:
         | undefined