+++ CHANGELOG.md
@@ -8,16 +8,48 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-### Fixes
+-   Added `ThirdPartyPasswordless` recipe + tests
 
+## [9.0.0] - 2022-02-20
+
+### Breaking Change
+
+-   Adds user context to all functions exposed to the user, and to API and Recipe interface functions. This is a non breaking change for User exposed function calls, but a breaking change if you are using the Recipe or APIs override feature
+-   Returns session from API interface functions that create a session
+-   Renames functions in ThirdPartyEmailPassword recipe (https://github.com/supertokens/supertokens-node/issues/219):
+    -   Recipe Interface:
+        -   `signInUp` -> `thirdPartySignInUp`
+        -   `signUp` -> `emailPasswordSignUp`
+        -   `signIn` -> `emailPasswordSignIn`
+    -   API Interface:
+        -   `emailExistsGET` -> `emailPasswordEmailExistsGET`
+    -   User exposed functions (in `recipe/thirdpartyemailpassword/index.ts`)
+        -   `signInUp` -> `thirdPartySignInUp`
+        -   `signUp` -> `emailPasswordSignUp`
+        -   `signIn` -> `emailPasswordSignIn`
+
+### Change:
+
+-   Uses recipe interface inside session class so that any modification to those get reflected in the session class functions too.
+
+## [8.6.1] - 2022-02-09
+
+### Refactor
+
+-   Removes unused property from session recipe
+
+## [8.6.0] - 2022-01-31
+
+### Changed
+
+-   Added userId as an optional property to the response of `recipe/user/password/reset` (compatibility with CDI 2.12).
 -   Fixes https://github.com/supertokens/supertokens-node/issues/244 - throws an error if a user tries to update email / password of a third party login user.
--   Adds thirdpartypasswordless recipe
 
 ### Added
 
+-   Adds ability to give a path for each of the hostnames in the connectionURI: https://github.com/supertokens/supertokens-node/issues/252
 -   add workflow to verify if pr title follows conventional commits
 -   Added `regenerateAccessToken` as a new recipe function for the session recipe.
--   Adds `ThirdPartyPasswordless` tests
 
 ### Breaking changes: